refactor: ListNewsstand에서 MainNews 분리
