+++ newsstand/styles/newsstand.css
@@ -107,7 +107,7 @@
     display: flex;
     flex-direction: column;
     height: auto;
-    width: 40%;
+    width: 100%;
     padding: 2%;
 }
 .list-news-left-top {
@@ -124,7 +124,7 @@
     justify-content: center;
     align-items: center;
     width: 100%;
-    min-height: 50%;
+    min-height: 210px;
 }
 
 .list-news-left-bottom {
